server: move newPodNetwork to a more logical place
